Chapter slots refactor. Removed mixin and flattened to individual slots components
