+++ components/chapter-slots/playfair.vue
@@ -21,20 +21,57 @@
 </template>
 
 <script>
-import ChapterSlots from "@/components/mixins/ChapterSlots";
 import DifferenceVisual from "@/components/chapters/playfair/DifferenceVisual.vue";
 import PChart from "@/components/chapters/playfair/PChart.vue";
 import BySide from "@/components/chapters/playfair/BySide.vue";
 import HoverText from "../global/HoverText";
+import DocsRenderer from "../docs-renderer/DocsRenderer";
 
 export default {
+  props: {
+    docContent: Array,
+  },
+  created() {
+    this.$store.commit("currentChapter/initializeChapterState", {
+      initialState: this.$options.chapterState,
+    });
+  },
+  mounted() {},
+  computed: {
+    chapterState() {
+      const obj = {};
+      Object.keys(this.$options.chapterState).forEach((key) =>
+        Object.defineProperty(obj, key, {
+          get: () => this.$store.state.currentChapter.chapterState[key],
+          set: (value) => {
+            this.$store.commit("currentChapter/updateChapterState", {
+              key,
+              value,
+            });
+          },
+          enumerable: true,
+        })
+      );
+      return obj;
+    },
+  },
   components: {
+    Slots: {
+      render(createElement) {
+        return createElement(DocsRenderer, {
+          props: {
+            docContent: this.$parent.docContent,
+          },
+          scopedSlots: this.$scopedSlots,
+          on: this.$parent.$listeners,
+        });
+      },
+    },
     DifferenceVisual,
     PChart,
     BySide,
     HoverText,
   },
-  mixins: [ChapterSlots],
   data() {
     return {
       slots: {
