+++ components/page/chapter/ChapterLayout.vue
@@ -47,6 +47,9 @@ export default {
     config: {
       type: Object,
     },
+    chapterSlot: {
+      type: Object,
+    },
   },
   components: {
     ChapterContent,
@@ -59,7 +62,7 @@ export default {
   provide() {
     return {
       theme: this.theme,
-      docsRenderer: this.docsRendererComponent,
+      chapterSlot: this.chapterSlot,
       docsComponents: this.docsComponents,
       componentData: this.componentData,
     };
@@ -105,14 +108,14 @@ export default {
     });
   },
   computed: {
-    docsRendererComponent() {
-      if (this.config.id) {
-        // https://vuejs.org/v2/guide/components-dynamic-async.html#Async-Components
-        // Loads the correct slots component from the chapter-slots folder. That component
-        // handles the doc rendering, injecting the specified slots into the docs renderer.
-        return () => import(`@/components/chapter-slots/${this.config.id}`);
-      }
-    },
+    // docsRendererComponent() {
+    //   if (this.config.id) {
+    //     // https://vuejs.org/v2/guide/components-dynamic-async.html#Async-Components
+    //     // Loads the correct slots component from the chapter-slots folder. That component
+    //     // handles the doc rendering, injecting the specified slots into the docs renderer.
+    //     return () => import(`@/components/chapter-slots/${this.config.id}`);
+    //   }
+    // },
     theme() {
       return this.config.theme;
     },
