+++ pages/chapters/dubois.vue
@@ -4,6 +4,7 @@
       v-if="docsData"
       :docData="docsData"
       :config="config"
+      :chapterSlot="duboisSlot"
     ></ChapterLayout>
   </div>
 </template>
@@ -12,6 +13,7 @@
 import ChapterLayout from "~/components/page/chapter/ChapterLayout";
 import chaptersConfig from "~/chaptersConfig";
 import docsData from "~/docs/dubois.json";
+import duboisSlot from "~/components/chapter-slots/dubois";
 
 export default {
   data() {
@@ -20,6 +22,7 @@ export default {
     return {
       docsData,
       config,
+      duboisSlot,
     };
   },
   computed: {},
