+++ pages/chapters/sandbox.vue
@@ -4,6 +4,7 @@
       v-if="docsData"
       :docData="docsData"
       :config="config"
+      :chapterSlot="sandboxSlot"
     ></ChapterLayout>
   </div>
 </template>
@@ -12,6 +13,7 @@
 import ChapterLayout from "~/components/page/chapter/ChapterLayout";
 import chaptersConfig from "~/chaptersConfig";
 import docsData from "~/docs/sandbox.json";
+import sandboxSlot from "~/components/chapter-slots/sandbox";
 
 export default {
   data() {
@@ -20,6 +22,7 @@ export default {
     return {
       docsData,
       config,
+      sandboxSlot,
     };
   },
   computed: {},
